fix(v6): the frozen-prose box is the pen's extent, not the declared cell's
`session::v6_screen_model` synthesises a box for the retired ("frozen") prose
layer, and its own comment says what that box is: "the frozen prose's own
extent, not the window's". It measured `chars * cell.w` — `V6Cell::run_px`,
documented in `zvm` as "uniform on purpose, even for a machine that painted
proportionally". Right for a box a GAME reserved; wrong for an extent recovered
from runs, and a retired entry has no game-declared box to inherit.

It matters because `build_chrome_canvas` turns that `w_px` into the
proportional pen's CLIP BOUND and floods `fill_explicit_bg_rows` to the same
short width. On Arthur's Amiga press the face advances ~10.4 native px against
a declared 8, so the box under-measured its own runs by about a quarter and the
tail of every frozen line fell outside the bound. Every fixed-pen machine is
byte-identical, because `V6Metric::advance` answers `cell.w` for every style
there — which is why this outlived the SQ-1054 sweep.

The two facts have to meet — a proportional pen AND retired prose — and no
medium in `stories/` puts them together: measured across the corpus, Arthur's
Amiga floppy has the only proportional release face and produces no retired
runs on any route driven here, Shogun's Amiga floppy produces nine and its face
is fixed, and the Macintosh presses are proportional only with a System disk
the repo cannot carry. The model is a pure function of the windows and the
metric, so the case boots the real frame that HAS the runs and hands it the
pen. Falsified at 432 px against the pen's 555.

Completes: SQ-1062

+++ crates/app/src/render/screen.rs
@@ -8098,7 +8098,7 @@ mod tests {
         let (fw, fh) = (u32::from(cell.w), u32::from(cell.h));
         // Artwork opaque under the FOURTH glyph alone.
         const OVER: usize = 3;
-        let mut gfx = image::RgbaImage::new(fw as u32 * 8, fh);
+        let mut gfx = image::RgbaImage::new(fw * 8, fh);
         for y in 0..fh {
             for x in (OVER as u32 * fw)..((OVER as u32 + 1) * fw) {
                 gfx.put_pixel(x, y, image::Rgba([9, 9, 9, 255]));
